refactor: 리뷰 내용 적용
 - 예약 시간, 테마 삭제시 예약이 존재하는 지 확인 후 삭제하도록 변경(예약이 있으면 삭제 불가하도록 Table 변경)
 - GlobalExceptionHandler가 좀더 명확하게 처리하도록 변경
    - 해당 변경점을 위한 테스트 추가

+++ src/main/java/roomescape/apply/reservationtime/application/ReservationTimeDeleter.java
@@ -1,6 +1,8 @@
 package roomescape.apply.reservationtime.application;
 
 import org.springframework.stereotype.Service;
+import roomescape.apply.reservation.application.ReservationFinder;
+import roomescape.apply.reservationtime.application.exception.ReservationTimeReferencedException;
 import roomescape.apply.reservationtime.domain.repository.ReservationTimeRepository;
 import roomescape.apply.reservationtime.application.exception.NotFoundReservationTimeException;
 
@@ -8,13 +10,20 @@ import roomescape.apply.reservationtime.application.exception.NotFoundReservatio
 public class ReservationTimeDeleter {
 
     private final ReservationTimeRepository reservationTimeRepository;
+    private final ReservationFinder reservationFinder;
 
-    public ReservationTimeDeleter(ReservationTimeRepository reservationTimeRepository) {
+    public ReservationTimeDeleter(ReservationTimeRepository reservationTimeRepository, ReservationFinder reservationFinder) {
         this.reservationTimeRepository = reservationTimeRepository;
+        this.reservationFinder = reservationFinder;
     }
 
     public void deleteReservationTimeBy(long id) {
         final long existId = reservationTimeRepository.checkIdExists(id).orElseThrow(NotFoundReservationTimeException::new);
+        boolean isReferenced = reservationFinder.findAnyByTimeId(id).isPresent();
+        if (isReferenced) {
+            throw new ReservationTimeReferencedException();
+        }
+
         reservationTimeRepository.deleteById(existId);
     }
 }
